+++ Editor/YarnImporterEditor.cs
@@ -20,10 +20,10 @@ public class YarnImporterEditor : ScriptedImporterEditor
     private SerializedProperty compilationErrorMessageProperty;
     private SerializedProperty localizationsProperty;
 
-    public string DestinationProgramError => destinationYarnProgramImporter?.compileError ?? null;
+    public string DestinationProgramError => destinationYarnProjectImporter?.compileError ?? null;
 
-    private YarnProgram destinationYarnProgram;
-    private YarnProgramImporter destinationYarnProgramImporter;
+    private YarnProject destinationYarnPrject;
+    private YarnProjectImporter destinationYarnProjectImporter;
 
     public override void OnEnable()
     {
@@ -41,11 +41,11 @@ public class YarnImporterEditor : ScriptedImporterEditor
 
     private void UpdateDestinationProgram()
     {
-        destinationYarnProgram = (target as YarnImporter).DestinationProgram;
+        destinationYarnPrject = (target as YarnImporter).DestinationProject;
 
-        if (destinationYarnProgram != null)
+        if (destinationYarnPrject != null)
         {
-            destinationYarnProgramImporter = AssetImporter.GetAtPath(AssetDatabase.GetAssetPath(destinationYarnProgram)) as YarnProgramImporter;
+            destinationYarnProjectImporter = AssetImporter.GetAtPath(AssetDatabase.GetAssetPath(destinationYarnPrject)) as YarnProjectImporter;
         }
     }
 
@@ -71,16 +71,16 @@ public class YarnImporterEditor : ScriptedImporterEditor
 
         EditorGUILayout.PropertyField(baseLanguageIdProperty);
 
-        if (destinationYarnProgram == null) {
-            EditorGUILayout.HelpBox("This script is not currently part of a Yarn Program, so it can't be compiled or loaded into a Dialogue Runner. Either click Create New Yarn Program, or Assign to Existing Yarn Program.", MessageType.Info);
-            if (GUILayout.Button("Create New Yarn Program...")) {
-                YarnImporterUtility.CreateYarnProgram(target as YarnImporter);
+        if (destinationYarnPrject == null) {
+            EditorGUILayout.HelpBox("This script is not currently part of a Yarn Project, so it can't be compiled or loaded into a Dialogue Runner. Either click Create New Yarn Project, or Assign to Existing Yarn Project.", MessageType.Info);
+            if (GUILayout.Button("Create New Yarn Project...")) {
+                YarnImporterUtility.CreateYarnProject(target as YarnImporter);
                 
                 UpdateDestinationProgram();
 
             }
             if (GUILayout.Button("Assign to Existing Yarn Program...")) {
-                var programPath = EditorUtility.OpenFilePanelWithFilters("Select an existing Yarn Program", Application.dataPath, new string[] {"Yarn Program (.yarnprogram)", "yarnprogram"} );
+                var programPath = EditorUtility.OpenFilePanelWithFilters("Select an existing Yarn Project", Application.dataPath, new string[] {"Yarn Program (.yarnprogram)", "yarnprogram"} );
                 programPath = "Assets" + programPath.Substring( Application.dataPath.Length );
                 if ( !string.IsNullOrEmpty(programPath) ) {
                     YarnImporterUtility.AssignScriptToProgram( (target as YarnImporter).assetPath, programPath);
@@ -90,7 +90,7 @@ public class YarnImporterEditor : ScriptedImporterEditor
             
         } else {
             using (new EditorGUI.DisabledGroupScope(true)) {
-                EditorGUILayout.ObjectField("Program", destinationYarnProgram, typeof(YarnProgramImporter), false);
+                EditorGUILayout.ObjectField("Program", destinationYarnPrject, typeof(YarnProjectImporter), false);
             }
         }
 
@@ -181,10 +181,10 @@ public class YarnImporterEditor : ScriptedImporterEditor
     {
         // First, gather all existing line tags, so that we don't
         // accidentally overwrite an existing one. Do this by finding _all_
-        // YarnPrograms, and by extension their importers, and get the
+        // YarnProjects, and by extension their importers, and get the
         // string tags that they found.
 
-        var allLineTags = Directory.GetFiles(Directory.GetCurrentDirectory(), "*.yarn") // get all yarn programs 
+        var allLineTags = Directory.GetFiles(Directory.GetCurrentDirectory(), "*.yarn") // get all yarn scripts 
             .Select(path => AssetImporter.GetAtPath(path)) // get the asset importer for that path
             .OfType<YarnImporter>() // ensure that they're all YarnImporters
             .SelectMany(importer => importer.stringIDs) // Get all of the string IDs in the base localization
@@ -247,7 +247,7 @@ public class YarnImporterEditor : ScriptedImporterEditor
                     {
                         var guid = AssetDatabase.AssetPathToGUID(importer.assetPath);
                         
-                        previousLocalizationDatabase.RemoveTrackedProgram(guid);
+                        previousLocalizationDatabase.RemoveTrackedProject(guid);
                         
                         // Mark that the localization database has changed,
                         // so needs to be saved
@@ -261,7 +261,7 @@ public class YarnImporterEditor : ScriptedImporterEditor
                     foreach (YarnImporter importer in serializedObject.targetObjects)
                     {
                         var guid = AssetDatabase.AssetPathToGUID(importer.assetPath);
-                        database.AddTrackedProgram(guid);
+                        database.AddTrackedProject(guid);
 
                         // Mark that the localization database should save
                         // changes