+++ Samples~/Space/Scripts/NodeVisitedTracker.cs
@@ -17,9 +17,14 @@ namespace Yarn.Unity.Example
         [SerializeField] Yarn.Unity.DialogueRunner dialogueRunner;
 #pragma warning restore 0649
 
+        // a HashSet is like a List or Array except it can't guarantee the order of items 
+        // which is OK if you don't care about order and just want to search inside it, like we do here!
+        // but you could replace this List<string> if you wanted to log the precise order of visited nodes, etc.
         private HashSet<string> _visitedNodes = new HashSet<string>();
 
-        void Start()
+        // we call AddFunction in Awake instead of Start, to ensure the Yarn function "visited()" 
+        // is added to the Dialogue Runner before it starts automatically (if configured to start automatically)
+        void Awake()
         {
             // Register a function on startup called "visited" that lets
             // Yarn scripts query to see if a node has been run before.